requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
importlib-metadata==3.3.0
Markdown==3.3.3
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
